LayoutTests/imported/w3c/web-platform-tests/xhr/send-authentication-basic-setrequestheader-existing-session-expected.txt


FAIL XMLHttpRequest: send() - "Basic" authenticated request using setRequestHeader() when there is an existing session assert_equals: expected "26f79fa9-8557-4973-970f-d47d5d7143c4\npass" but got "1e77b02b-c11b-41f8-a81e-9f239b304736\nopen-pass"
Note: this test will only work as expected once per browsing session. Restart browser to re-test.
